requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

pydantic
pydantic-settings
python-dotenv
requests